"pandas ecosystem 2019" talk updated
While the content will be exactly the same, I'll be making my talk a bit more a story, based on feedback from PyConX (where it was also delivered). Because of this, the title "Breaking pandas" will be a bit more appropriate (and hopefully cooler) than "pandas ecosystem 2019". Updating the program with the proposed change.